requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
async-timeout==4.0.3
attrs==25.4.0
Babel==2.9.1
certifi==2026.2.25
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.3.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
magic-filter==1.0.12
mistralai==2.0.2
multidict==6.7.1
openai==2.26.0
opentelemetry-api==1.39.1
opentelemetry-semantic-conventions==0.60b1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
yarl==1.23.0
zipp==3.23.0
